MINOR: Stricter types

+++ src/Dev/Install/InstallRequirements.php
@@ -624,16 +624,16 @@ class InstallRequirements
 
         switch (strtolower(substr($memString, -1))) {
             case "k":
-                return round(substr($memString, 0, -1) * 1024);
+                return round((float)substr($memString, 0, -1) * 1024);
 
             case "m":
-                return round(substr($memString, 0, -1) * 1024 * 1024);
+                return round((float)substr($memString, 0, -1) * 1024 * 1024);
 
             case "g":
-                return round(substr($memString, 0, -1) * 1024 * 1024 * 1024);
+                return round((float)substr($memString, 0, -1) * 1024 * 1024 * 1024);
 
             default:
-                return round($memString);
+                return round((float)$memString);
         }
     }
